+++ tests/unit/test_state.py
@@ -27,6 +27,8 @@ def test_adapter_state_from_charm():
         "retry-count": 1,
         "retry-interval": 10,
         "retry-redispatch": True,
+        "paths": "/api/v1,/api/v2",
+        "subdomains": "api",
     }
     ingress_relation_data = IngressRequirerData(
         app=IngressRequirerAppData(model="model", name="name", port=8080),
@@ -40,6 +42,8 @@ def test_adapter_state_from_charm():
     assert charm_state.retry_count == charm.config.get("retry-count")
     assert charm_state.retry_interval == charm.config.get("retry-interval")
     assert charm_state.retry_redispatch == charm.config.get("retry-redispatch")
+    assert charm_state.paths == charm.config.get("paths").split(",")
+    assert charm_state.subdomains == charm.config.get("subdomains").split(",")
 
 
 def test_integrator_state_from_charm():
@@ -95,6 +99,22 @@ def test_state_from_charm_invalid_address():
         state.State.from_charm(charm, None)
 
 
+def test_state_from_charm_invalid_paths():
+    """
+    arrange: mock a charm with backend addresses, ports configuration and invalid paths
+    act: instantiate a State
+    assert: a InvalidStateError is raised
+    """
+    charm = Mock(CharmBase)
+    charm.config = {
+        "backend-addresses": "127.0.0.1,127.0.0.2",
+        "backend-ports": "8080,8081",
+        "paths": "invalid path",
+    }
+    with pytest.raises(state.InvalidStateError):
+        state.State.from_charm(charm, None)
+
+
 def test_state_from_charm_invalid_port():
     """
     arrange: mock a charm with backend address and without port configuration
@@ -133,6 +153,22 @@ def test_state_from_charm_invalid_retry_interval():
     charm = Mock(CharmBase)
     charm.config = {
         "retry-interval": -1,
+    }
+    with pytest.raises(state.InvalidStateError):
+        state.State.from_charm(charm, None)
+
+
+def test_state_from_charm_invalid_subdomains():
+    """
+    arrange: mock a charm with backend addresses, ports configuration and invalid subdomains
+    act: instantiate a State
+    assert: a InvalidStateError is raised
+    """
+    charm = Mock(CharmBase)
+    charm.config = {
+        "backend-addresses": "127.0.0.1,127.0.0.2",
+        "backend-ports": "8080,8081",
+        "subdomains": "invalid$subdomains",
     }
     with pytest.raises(state.InvalidStateError):
         state.State.from_charm(charm, None)